+++ changelog.md
@@ -54,6 +54,7 @@ This is a major update that breaks backwards compatibility.
 * Replaced all use of MD5 and SHA1 hash functions with SHA256.
 * Now checks for invalid host strings when sending via SMTP.
 * Include timestamps in HTML-format debug output
+* More reliable folding of message headers
 
 ## Version 5.2.23 (March 15th 2017)
 * Improve trapping of TLS errors during connection so that they don't cause warnings, and are reported better in debug output